Converged maven wrapper, fixed docker and compile fail
Some cleanups, including fixing the maven wrapper generation and
deleting orphaned wrapper classes. Deleted unused injection param of
DataSource, which broke docker-compose when MySQL isn't used. Switched
ZipkinServerProperties to be a top-level class as it had class
resolution issues finding its annotation.
