--- maven/maven-wrapper.properties
@@ -1,3 +1,0 @@
-#Maven download properties
-#Tue Sep 29 23:57:08 PDT 2015
-distributionUrl=https\://repository.apache.org/content/repositories/releases/org/apache/maven/apache-maven/3.3.3/apache-maven-3.3.3-bin.zip
